Add formatter for CalendarAlarm duration
Need to be able to convert from \DateInterval to string.

Oddly although the RFC 5545 duration string format is the same one that
DateInterval takes for its constructor, the class doesn't have a method
to return that format.

Used some of the advice in http://sixrevisions.com/web-development/php-dateinterval-class/ to build this solution.