Fix $item cutting in CalendarStream->addItem()
When the mb_strcut happens to cut through a multibyte character at the end of the cutting length, it ommits the leading byte of that character resulting in less than self::LINE_LENGTH bytes added to the $block. Number added to $start should not be constant. Next cycle, the next mb_strcut also cuts through the same multibyte character, this time including it and doing a shift back in the $item. The next cycle, one single byte character is ommited in the output.

To recreate this behavoir, create $item like this:
 69 single byte characters
 1 multibyte character
 >= 72 single byte characters (to see the character missing)